refactor: migrate package inside
