+++ src/main/java/kitchenpos/repository/OrderTableRepository.java
@@ -4,6 +4,7 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 import kitchenpos.domain.OrderTable;
+import kitchenpos.repository.entity.OrderTableEntityRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.Repository;
 import org.springframework.data.repository.query.Param;
