+++ src/main/java/kitchenpos/repository/TableGroupRepository.java
@@ -3,6 +3,7 @@ package kitchenpos.repository;
 import java.util.List;
 import java.util.Optional;
 import kitchenpos.domain.TableGroup;
+import kitchenpos.repository.entity.TableGroupEntityRepository;
 import org.springframework.data.repository.Repository;
 
 public interface TableGroupRepository extends Repository<TableGroup, Long>, TableGroupEntityRepository {
