+++ src/main/java/kitchenpos/repository/entity/MenuEntityRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.repository;
+package kitchenpos.repository.entity;
 
 import java.util.Collection;
