+++ src/main/java/kitchenpos/repository/entity/OrderEntityRepositoryImpl.java
@@ -1,6 +1,7 @@
-package kitchenpos.repository;
+package kitchenpos.repository.entity;
 
 import kitchenpos.domain.Order;
+import kitchenpos.repository.OrderRepository;
 import org.springframework.context.annotation.Lazy;
 
 public class OrderEntityRepositoryImpl implements OrderEntityRepository {
