+++ src/main/java/kitchenpos/repository/entity/OrderTableEntityRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.repository;
+package kitchenpos.repository.entity;
 
 import java.util.Collection;
 import java.util.List;
