+++ src/main/java/kitchenpos/repository/entity/TableGroupEntityRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.repository;
+package kitchenpos.repository.entity;
 
 import kitchenpos.domain.TableGroup;
